django-server/skyjacs_server/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
pycodestyle==2.3.1
pyflakes==1.5.0
pytz==2018.3
Pillow==4.2.1
fuzzywuzzy==0.16.0
